A Snake Game
I had coded the game from a youtube creator "BroCode". This game was actually much more difficult to understand in terms of what was going on in the code. I was able to learn a lot about how to use array's in different ways in order to make an object move in a particular way.